requirements.txt
gTTS==2.5.1
langchain==0.2.14
langchain_core==0.2.32
langchain_groq==0.1.9
numpy==2.0.1
openai_whisper==20231117
PyAudio==0.2.14
pygame==2.5.2
python-decouple==3.8
python-dotenv==1.0.1
scipy==1.14.0
streamlit==1.34.0
fastapi
